\name{lolaRomero-package}
\alias{lolaRomero-package}
\alias{lolaRomero}
\docType{package}
\title{
\packageTitle{lolaRomero}
}
\description{
\packageDescription{lolaRomero}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lolaRomero}
\packageIndices{lolaRomero}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lolaRomero}

Maintainer: \packageMaintainer{lolaRomero}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
