\name{lolaRomeroPACKAGE-package}
\alias{lolaRomeroPACKAGE-package}
\alias{lolaRomeroPACKAGE}
\docType{package}
\title{
\packageTitle{lolaRomeroPACKAGE}
}
\description{
\packageDescription{lolaRomeroPACKAGE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lolaRomeroPACKAGE}
\packageIndices{lolaRomeroPACKAGE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lolaRomeroPACKAGE}

Maintainer: \packageMaintainer{lolaRomeroPACKAGE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
